--- matjalalzz/src/main/java/shop/matjalalzz/domain/party/entity/GenderCondition.java
@@ -1,5 +1,0 @@
-package shop.matjalalzz.domain.party.entity;
-
-public enum GenderCondition {
-	M,W,A
-}
