--- matjalalzz/src/main/java/shop/matjalalzz/domain/party/entity/PartyStatus.java
@@ -1,5 +1,0 @@
-package shop.matjalalzz.domain.party.entity;
-
-public enum PartyStatus {
-	RECRUITING, COMPLETED, TERMINATED
-}
