+++ matjalalzz/src/main/java/shop/matjalalzz/party/dao/PartyRepository.java
@@ -1,7 +1,8 @@
-package shop.matjalalzz.domain.party.dao;
+package shop.matjalalzz.party.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.matjalalzz.domain.party.entity.Party;
+import shop.matjalalzz.party.entity.Party;
 
 public interface PartyRepository extends JpaRepository<Party, Long> {
+
 }
