+++ matjalalzz/src/main/java/shop/matjalalzz/party/dto/PartyDetailResponse.java
@@ -1,14 +1,15 @@
-package shop.matjalalzz.domain.party.dto;
+package shop.matjalalzz.party.dto;
 
 import java.time.LocalDateTime;
 import lombok.Builder;
-import shop.matjalalzz.domain.party.entity.GenderCondition;
-import shop.matjalalzz.domain.party.entity.PartyStatus;
+import shop.matjalalzz.party.entity.GenderCondition;
+import shop.matjalalzz.party.entity.PartyStatus;
 
 @Builder
-public record PartyListResponse(
+public record PartyDetailResponse(
     Long partyId,
     String title,
+    String description,
     PartyStatus status,
     int maxCount,
     int minCount,
@@ -19,9 +20,7 @@ public record PartyListResponse(
     LocalDateTime metAt,
     LocalDateTime deadline,
     LocalDateTime createdAt,
-    Long hostId,
-    String shopName,
-    String shopAddress
+    Long hostId
 ) {
 
 }
