+++ matjalalzz/src/main/java/shop/matjalalzz/party/entity/GenderCondition.java
@@ -1,0 +1,5 @@
+package shop.matjalalzz.party.entity;
+
+public enum GenderCondition {
+    M, W, A
+}
