+++ matjalalzz/src/main/java/shop/matjalalzz/party/entity/PartyStatus.java
@@ -1,0 +1,5 @@
+package shop.matjalalzz.party.entity;
+
+public enum PartyStatus {
+    RECRUITING, COMPLETED, TERMINATED
+}
