+++ matjalalzz/src/main/java/shop/matjalalzz/party/mock/dao/MockUserRepository.java
@@ -1,7 +1,7 @@
-package shop.matjalalzz.domain.party.mock.dao;
+package shop.matjalalzz.party.mock.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.matjalalzz.domain.party.mock.entity.MockUser;
+import shop.matjalalzz.party.mock.entity.MockUser;
 
 public interface MockUserRepository extends JpaRepository<MockUser, Long> {
